+++ src/GlobalRenderer.js
@@ -1,7 +1,6 @@
 import React, { Suspense, Fragment, useLayoutEffect, useRef } from 'react'
 import PropTypes from 'prop-types'
 import { useThree, useFrame } from 'react-three-fiber'
-import { sRGBEncoding, NoToneMapping } from 'three'
 
 import config from './config'
 import { useCanvasStore } from './store'
@@ -12,37 +11,29 @@ import { useScrollRig } from './useScrollRig'
  */
 const GlobalRenderer = ({ children }) => {
   const scene = useRef()
-  const { gl, size } = useThree()
+  const { gl } = useThree()
   const canvasChildren = useCanvasStore((state) => state.canvasChildren)
   const scrollRig = useScrollRig()
 
   useLayoutEffect(() => {
-    gl.outputEncoding = sRGBEncoding
-    gl.setClearColor(0x000000, 0)
     gl.debug.checkShaderErrors = config.debug
-    gl.toneMapping = NoToneMapping
   }, [])
 
-  // GLOBAL RENDER LOOP
+  // PRELOAD RENDER LOOP
   useFrame(({ camera, scene }) => {
+    gl.autoClear = false
     // Render preload frames first and clear directly
     config.preloadQueue.forEach((render) => render(gl))
     if (config.preloadQueue.length) gl.clear()
+    // cleanup
+    config.preloadQueue = []
+  }, config.PRIORITY_PRELOAD)
 
-    // prevent viewport from being cleared if we have multiple render passes
-    if (config.viewportQueueBefore.length || config.viewportQueueAfter.length || config.scissorQueue.length) {
-      gl.autoClear = false
-    }
-
-    // Render viewports passes before
-    if (config.viewportQueueBefore.length) {
-      config.viewportQueueBefore.forEach((render) => render(gl, size))
-    }
-
+  // GLOBAL RENDER LOOP
+  useFrame(({ camera, scene }) => {
     // Global render pass
     if (config.globalRender) {
-      // run any pre-process frames
-      config.preRender.forEach((render) => render(gl))
+      gl.autoClear = false // will fail in VR
 
       // render default layer, scene, camera
       camera.layers.disableAll()
@@ -52,30 +43,11 @@ const GlobalRenderer = ({ children }) => {
       gl.clearDepth() // render as HUD over any other renders
       gl.render(scene, camera)
 
-      // run any post-render frame (additional layers etc)
-      config.postRender.forEach((render) => render(gl))
-    }
+      // cleanup for next frame
+      config.globalRender = false
 
-    // Render global scissors
-    if (config.scissorQueue.length) {
-      config.scissorQueue.forEach((render) => render(gl, camera))
+      gl.autoClear = true
     }
-
-    // Render viewports after
-    if (config.viewportQueueAfter.length) {
-      gl.autoClear = false
-      config.viewportQueueAfter.forEach((render) => render(gl, size))
-    }
-
-    // cleanup for next frame
-    config.globalRender = false
-    config.preRender = []
-    config.postRender = []
-    config.preloadQueue = []
-    config.scissorQueue = []
-    config.viewportQueueBefore = []
-    config.viewportQueueAfter = []
-    gl.autoClear = true
   }, config.PRIORITY_GLOBAL) // Take over rendering
 
   config.debug && console.log('GlobalRenderer', Object.keys(canvasChildren).length)